requirements.txt
Django==1.11
django-braces==1.12.0
django-friendship==1.5.0
django-messages==0.5.3
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.7
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
